maint: allow external_nodes script to be specified
If you don't specify it, then external nodes aren't used.  This setting
should come from the puppet.conf, but since this setting goes in the
master section of that file and run_mode can't be set and defaults to
'user' not 'master', it doesn't get used.  So we have specify it
manually for now.