(#8356) Specify setting type for color
Because we default the color setting to "false" on Microsoft Windows,
the heuristics used to detect which type of setting we're using were
getting confused, and mis-detected color as being a BooleanSetting
rather than just a Setting.

By specifying that color is a "Setting", we can skip the
auto-detection, and avoid this problem entirely.

Reviewed-by: Josh Cooper <josh@puppetlabs.com>