(#8341) Do not explicitly load all facts looking for macosx_productversion_major
It used to be that Puppet would need to explicitly load all facts when
it was trying to use a fact that didn't have the same name as the file
it was defined in.

Facter has been fixed to automatically load all facts when it can't
find the one you're looking for by mapping the fact name to a file
name, so doing this explicitly outside of Facter is no longer
necessary.

By removing the extra explicit loading of all Facter facts in Puppet,
we now require Facter >= 1.6.0 (2255abe).