Fix environment validate_dirs bug
The environment validates its modulepath and manifestdir settings, but Dir.getwd
File.join where creating a path with both back and forward slashes. The call
to expand_path normalizes it to all forward slashes.

But as a result of this change it exposed a bug in our rdoc parser, whereby
it attempted to ensure the module path was an ancestor of the directory
containing the module we want to generate documentation from. But it assumed
the root was '/', which caused an infinite loop when passed Windows paths.

I added a test for this case, which is only run on Windows, because on Unix
File.dirname("C:/") == '.'.

I had to disable one rdoc spec test, because it attempts to reproduce a
specific bug and in the process creates a path of the form
C:/.../files/C:/..., which fails because ':' is not a valid filename
character on Windows (excluding the prefix).