Update tests for Windows
The autoload specs now pass on Windows.

The inventory spec was failing because Time.now.strftime("%Z")
returns "Pacific Daylight Time" on Windows, instead of "PDT" on
other platforms. As a result, the split method was sometimes
splitting in the wrong places.

This could be a problem if a Windows agent every formatted a
timezone and sent that to the puppetmaster, and the
puppetmaster assumed the timezone did not contain spaces.

As far as I could tell, the only place the inventory cert is
formatted and then later parsed, is on the puppetmaster itself,
when keeping track of certs that it has issued. If this is
incorrect, then the cert format method will probably need to quote
the timezone, or investigate other ways to format the timezone,
such as the UTC offset.