Treat Windows absolute paths as absolute paths
Previously, we only considered files that matched the *nix concept of
'absolute' as being absolute paths.  Since absolute paths on Windows
look more like URLs with this world-view, we need to specifically look
for the Windows absolute paths, and treat them as such.

We will still treat *nix absolute paths as absolute on Windows, even
though they are actually relative to the "current" drive.  We do not
currently limit which "style" of absolute path is allowed based on
what the agent is.