(#3124) Seperate defined and declared function.
Currently defined checks both if a class/type is defined as well as
if a resource is declared. Instead of having both functionality mixed
in the same function, we have seperated them out to two seperate
functions. If defined is used to detect whether a resource is declared
a deprecation warning will be thrown.