Prototype partial solution to #6539
When instantiating a resource for a type, (and instantiating its parent type)
we check if its parent type has arguments. If it does, we select from our
parameters those which correspond to arguments of the parent type, and pass
them along. Similarly, we select those arguments from the parent type which we
don't already have, and add them as arguments to our own type.

Defaults aren't currently inherited properly in either direction (though it's
not clear whether they should be inherited up).