(#3234) Don't fail if a node can't be found in active record
Previously, storedconfigs would fail each time it tried to compile a new
node, as active_record was configured as the node cache, but improperly
handled missing nodes. It would try to find a node in the database and
then call fact_merge on it, but didn't correctly handle the case where
the node wasn't in the database (for instance, on a first compile). Now
it will check for that case and simply return nil, so the compile can
continue on to the 'plain' terminus, using a fresh Node instance.