Use Environment#module to find modules when looking for manifests
This was using Module.find and now instead uses Environment#module. This
means there will always be a module, so instead of including the code
for matching manifests outside of a module in this method, we define a
match_manifests method on NullModule.

NullModule is a subclass of Module now, with a singleton instance. It
has an implementation of match_manifests, and null implementations of
plugins, manifests, etc methods.