Properly track module info for classes loaded via absolute import
These were being implicitly considered as being inside the NullModule,
even though the absolute path may be inside of an existing module. Now
we still glob files using the NullModule, but defer choosing which
module to associate with the file until later when we are about to
evaluate the file.