Fix more failing tests
These were failing because they stub a lot of objects relating to the
scope and catalog, but that means the resource_type and module
information was missing. Also, we now enforce that resource_type must
always have a module, by providing a default (the null module) when none
is specified.