Add support for application instances to TypeCollection and Type
The TypeCollection now tracks application definitions in a new array
'applications'. Type recognizes application instances as a different kind of
resource, similar to how definitions and hostclasses are handled.

The one wrinkle with this is that we have the component/node mapping in
a special 'nodes' parameter in application instances. Type#valid_parameter?
has that knowledge hardcoded for now.