Simplify environment catalog format
Previously we were returning a structure like

{"Forge[production]" =>
  {"Web[w1]" => {...},
   "Web[w2]" => {...}}

This was cumbersome to work with because it used resource refs to refer
to resources, and particularly because it used those resource refs as
hash keys.

Now we refer to resource using a combination of :type/:title, so we
don't need to parse the refs later. This is also more consistent with
the structure of agent catalogs. The structure has also been flattened,
so that we represent an application (and each of its component) within a
single map, rather than spread out between the key and value of a
collection.

[{:type => "Forge",
  :title => "production",
  :components => [{:type => "Web", :title => "w1", ...},
                  {:type => "Web", :title => "w2", ...}]}]

This minimizes the amount of processing we need to do in order to
actually consume this data structure.