Add an 'export' metaparam
This adds the metaparam that controls whether a capability resource is
actually produced and exported or not.

Note that produced resources are always evaluated, since we need to find
out their type and title before we can tell whether they should have been
produced in the first place. If we determine that they shouldn't be
exported, we kill them again

Fixes https://tickets.puppetlabs.com/browse/APPMGMT-67