Puppet::Resource::Type (evaluate_produces): new method
Move evaluation of produced capability resources from evaluate_code into
its own method.

One consequence of this change is that capability resources are evaluated
in their own scope, separate from the body of the define that produces them
(since we aren't actually evaluating that define during environment catalog
compiler). During node compilation, that means the same scope will be
created twice, and therefore any variables set during evaluation of the
`produces` expression won't be visible to the body of the definition. In
practice this shouldn't be a problem, because the only way a variable could
be defined at that time is through a strange function call.